gestion_de_ordenes_de_compra-area_adquisiciones-main/OCApi/requirements.txt
Django==3.2.6
mysqlclient==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-cors-headers==4.2.0
requests==2.31.0
sendgrid==6.11.0
python-dotenv==1.0.1